[COREVM-105] Implement mechanism to version bytecode loader
