+++ include/frontend/bytecode_loader.h
@@ -1,7 +1,7 @@
 /*******************************************************************************
 The MIT License (MIT)
 
-Copyright (c) 2014 Yanzheng Li
+Copyright (c) 2015 Yanzheng Li
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
@@ -20,12 +20,15 @@ COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 *******************************************************************************/
-#ifndef COREVM_SCHEMA_REPOSITORY_H_
-#define COREVM_SCHEMA_REPOSITORY_H_
+#ifndef COREVM_FRONTEND_BYTECODE_LOADER_H_
+#define COREVM_FRONTEND_BYTECODE_LOADER_H_
 
-#include <set>
+#include "errors.h"
+#include "../runtime/process.h"
+
+#include <sneaker/json/json.h>
+
 #include <string>
-#include <unordered_map>
 
 
 namespace corevm {
@@ -34,13 +37,10 @@ namespace corevm {
 namespace frontend {
 
 
-class schema_repository {
-public:
-  static const std::string load_by_format_version(const std::string&);
-  static const std::set<std::string> list_supported_versions();
-private:
-  static const std::unordered_map<std::string, std::string> format_version_to_schema_map;
-};
+using sneaker::json::JSON;
+
+
+void load(const std::string&, corevm::runtime::process&) throw(corevm::frontend::file_loading_error);
 
 
 } /* end namespace frontend */
@@ -49,4 +49,4 @@ private:
 } /* end namespace corevm */
 
 
-#endif /* COREVM_SCHEMA_REPOSITORY_H_ */
+#endif /* COREVM_FRONTEND_BYTECODE_LOADER_H_ */
